[sandbox] extract sandbox from policy loader
* to split concerns of policy loading and a sandbox
